Patch mysterious penultimate state bug
The app either crashes or produces the incorrect
optimal move when a "penultimate" state is
reached.

By "penultimate" state, I mean a state that is
can lead to a terminal state in one action.

I'm still not sure what the cause is.
However, I manually searched for winning actions,
which seems to have fixed the issue.
Ideally, however, it would be nice to fix the
problem at the root.